# Conflicts:
#	shared/src/commonMain/kotlin/ui/board/BoardManagementScreen.kt
#	shared/src/iosMain/kotlin/background/BackgroundRefreshManager.kt
#	shared/src/iosMain/kotlin/ui/board/PlatformVideoPlayer.ios.kt

+++ shared/src/iosMain/kotlin/util/DevicePerformance.ios.kt
@@ -10,7 +10,7 @@ import platform.Foundation.NSProcessInfo
 @OptIn(ExperimentalForeignApi::class)
 public actual fun detectDevicePerformanceProfile(platformContext: Any?): DevicePerformanceProfile {
     val processInfo = NSProcessInfo.processInfo
-    val totalRamMb = (processInfo.physicalMemory / (1024 * 1024)).toInt()
+    val totalRamMb = (processInfo.physicalMemory.toLong() / (1024 * 1024)).toInt()
     // 2GB未満の端末を低RAM扱いにする
     val isLowRam = totalRamMb < 2048
